Transitioning supervisor to a superclass
