+++ lib/python/brand/__init__.py
@@ -2,7 +2,12 @@ from .tools import (get_node_parameter_value, get_parameter_value,
                          initializeRedisFromYAML, get_node_parameter_dump,
                          get_redis_info, main, get_node_io, unpack_string,
                          node_stage)
+
 from .node import BRANDNode 
+
+from .supervisor import Supervisor
+
+from .booter import Booter
 
 from .exceptions import (GraphError, NodeError, 
                         BooterError, DerivativeError,
