+++ lib/python/brand/booter.py
@@ -16,7 +16,7 @@ import traceback
 import coloredlogs
 import redis
 
-from brand import (GraphError, NodeError, CommandError)
+from .exceptions import (GraphError, NodeError, CommandError)
 
 DEFAULT_REDIS_IP = '127.0.0.1'
 DEFAULT_REDIS_PORT = 6379
@@ -328,48 +328,39 @@ class Booter():
         sys.exit(0)
 
 
-def parse_booter_args():
-    """
-    Parse command-line arguments for Booter
+    def parse_booter_args():
+        """
+        Parse command-line arguments for Booter
 
-    Returns
-    -------
-    args : Namespace
-        Booter arguments
-    """
-    ap = argparse.ArgumentParser()
-    ap.add_argument("-m",
-                    "--machine",
-                    required=True,
-                    type=str,
-                    help="machine on which this booter is running")
-    ap.add_argument("-i",
-                    "--host",
-                    required=False,
-                    type=str,
-                    default=DEFAULT_REDIS_IP,
-                    help="ip address of the redis server"
-                    f" (default: {DEFAULT_REDIS_IP})")
-    ap.add_argument("-p",
-                    "--port",
-                    required=False,
-                    type=int,
-                    default=DEFAULT_REDIS_PORT,
-                    help="port of the redis server"
-                    f" (default: {DEFAULT_REDIS_PORT})")
-    ap.add_argument("-l",
-                    "--log-level",
-                    default=logging.INFO,
-                    type=lambda x: getattr(logging, x),
-                    help="Configure the logging level")
-    args = ap.parse_args()
-    return args
-
-
-if __name__ == '__main__':
-    # parse command line arguments
-    args = parse_booter_args()
-    kwargs = vars(args)
-    # Run Booter
-    booter = Booter(**kwargs)
-    booter.run()
+        Returns
+        -------
+        args : Namespace
+            Booter arguments
+        """
+        ap = argparse.ArgumentParser()
+        ap.add_argument("-m",
+                        "--machine",
+                        required=True,
+                        type=str,
+                        help="machine on which this booter is running")
+        ap.add_argument("-i",
+                        "--host",
+                        required=False,
+                        type=str,
+                        default=DEFAULT_REDIS_IP,
+                        help="ip address of the redis server"
+                        f" (default: {DEFAULT_REDIS_IP})")
+        ap.add_argument("-p",
+                        "--port",
+                        required=False,
+                        type=int,
+                        default=DEFAULT_REDIS_PORT,
+                        help="port of the redis server"
+                        f" (default: {DEFAULT_REDIS_PORT})")
+        ap.add_argument("-l",
+                        "--log-level",
+                        default=logging.INFO,
+                        type=lambda x: getattr(logging, x),
+                        help="Configure the logging level")
+        args = ap.parse_args()
+        return args
